refactor: :recycle: change file name row by columns  within table desing folder in request page, so it can be more clear for user
